Create HW3
A simple code that demonstrates the architecture used which is a combination of design patterns such as Interpreter, Observer, and State patterns. The program is a simple Chat Bot that has a limited amount of responses and the user must input specific commands for the bot to understand. This is only a very basic and simple Chat Bot used mainly for testing and organize our project. More libraries will be added to the program as we add more functionalities to the Chat Bot.